==> session-5/example-2/1-generate_synthetic_queries.ipynb <==
# Generate Synthetic Queries with GPT-4o-mini (batch)

[Video link](https://youtu.be/hOLBrIjRAj4) | [Blog link](https://shawhin.medium.com/fine-tuning-text-embeddings-f913b882b11c)<br>
[Dataset](https://huggingface.co/datasets/shawhin/ai-job-embedding-finetuning) | [Fine-tuned Model](https://huggingface.co/shawhin/distilroberta-ai-job-embeddings) <br>
Based on example from [here](https://sbert.net/docs/sentence_transformer/training_overview.html#trainer)

### imports

import pandas as pd
import json
from datasets import load_dataset

import os
from openai import OpenAI
from dotenv import load_dotenv

# import sk from .env file
load_dotenv()

# connect to openai API
client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

### functions

prompt_template = lambda job_description : f"""Read the following job description and create a concise job search query with at most 3 specialized skills or \
areas of expertise that are distinct to the role. Exclude generic data science or software engineering skills like AI, machine \
learning, and coding languages unless they are explicitly highlighted as unique or advanced. Keep the query short and human-like, \
suitable for typing into a search engine. 

Here's the job description: {job_description}"""

def generate_query(job_description):
    """
        Function to generate synthetic query to input job description.
    """

    # generate prompt
    prompt = prompt_template(job_description)

    # make api call
    response = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[
            {"role": "user", "content": prompt}
        ], 
        temperature = 0.7
    )
    
    # return response
    return response.choices[0].message.content

### Load data

# load data from HF hub
ds = load_dataset("datastax/linkedin_job_listings")

# convert to pandas df
df = ds['train'].to_pandas()

# keep only title and description
df = df[['title', 'description']]
df.shape

# List of strings to search for
search_terms = ["Data Scientist", "Data Analyst", "Machine Learning Engineer", 
                "Data Engineer", "AI Engineer", "Deep Learning"]

# Create a regex pattern to match any of the strings
pattern = '|'.join(search_terms)

# Filter rows that contain any of the search terms
df = df[df['title'].str.contains(pattern, case=False, na=False)]
df.shape

# save to file
df.to_csv('data/job_data.csv')

### Generates queries for JDs

job_description_list = df['description'].to_list()

# # non-batch approach
# synthetic_query_list = []
# for job_description in job_description_list:
#     # generate synthetic query and append to list
#     synthetic_query_list.append(generate_query(job_description).replace('"',''))

# # add queries to df
# df['query'] = synthetic_query_list

# df.to_csv('data/job_data_w_query.csv')

#### 1) Create batch request file

# create batch requests
batch_requests = [
    {
        "custom_id": f"request-{i+1}",  # Custom ID for tracking
        "method": "POST",
        "url": "/v1/chat/completions",
        "body": {
            "model": "gpt-4o-mini",
            "messages": [
                {"role": "user", "content": prompt_template(job_description)}
            ],
            "temperature": 0.7
        }
    }
    for i, job_description in enumerate(job_description_list)
]

# Convert to JSONL format (newline-delimited JSON)
batch_jsonl = "\n".join(json.dumps(request) for request in batch_requests)

# Save to a .jsonl file
with open("data/batch_requests.jsonl", "w") as file:
    file.write(batch_jsonl)

#### 2) Upload batch request file

batch_input_file = client.files.create(
    file=open("data/batch_requests.jsonl", "rb"),
    purpose="batch"
)

print(batch_input_file)

#### 3) Create batch

# create batch job
batch_object = client.batches.create(
    input_file_id=batch_input_file.id,
    endpoint="/v1/chat/completions",
    completion_window="24h",
    metadata={
        "description": "synthetic queries from job descriptions"
    }
)

print(batch_object)
