==> session-5/example-2/4-use_finetuned_embeddings.ipynb <==
# Use Fine-tuned Model

Code authored by: Shaw Talebi

[Video link](https://youtu.be/hOLBrIjRAj4) | [Blog link](https://shawhin.medium.com/fine-tuning-text-embeddings-f913b882b11c)<br>
[Dataset](https://huggingface.co/datasets/shawhin/ai-job-embedding-finetuning) | [Fine-tuned Model](https://huggingface.co/shawhin/distilroberta-ai-job-embeddings) <br>
Based on example from [here](https://sbert.net/docs/sentence_transformer/training_overview.html#trainer)

### imports

from sentence_transformers import SentenceTransformer
from datasets import load_dataset
import numpy as np

### load model

model = SentenceTransformer("shawhin/distilroberta-ai-job-embeddings")

### load data

dataset = load_dataset("shawhin/ai-job-embedding-finetuning")

### use it!

# new query
query = "data scienist 6 year experience, LLMs, credit risk, content marketing"
query_embedding = model.encode(query)

# encode JDs
jd_embeddings = model.encode(dataset["test"]["job_description_pos"])

# compute similarities
similarities = model.similarity(query_embedding, jd_embeddings)
print(similarities.shape)

# sort JDs based on similarity
simiarities_argsorted = np.argsort(similarities.numpy(), axis=1)

# return most similar job description
dataset["test"]["job_description_pos"][int(simiarities_argsorted[0][-1])]
